fix: share dependency version parsing helpers
